# openapi.yaml
openapi: 3.0.3
info:
  title: DRP · Domestic Resource Planning API
  description: >
    API REST del core de DRP (Domestic Resource Planning).


    Este documento es la **fuente de verdad** de la API, no una ilustración
    de ella (ver ADR-007): se edita antes que el código que lo implementa, el
    frontend genera de él sus tipos y su cliente, y el backend se verifica
    contra él con las pruebas de contrato. Ver readme.md del proyecto para el
    contexto completo — modelo de dominio, event bus y roadmap.


    Dos convenciones transversales. Los **errores de forma** responden 400 con
    código `VALIDATION_ERROR`; los de **regla de negocio**, 409 con el código
    concreto. Y todas las colecciones se devuelven **paginadas**, con la misma
    envoltura `{ items, page, size, total }`.
  version: "0.3.0-fase2"
  license:
    name: Ver LICENSE del repositorio
servers:
  - url: http://localhost:8080/api/v1
    description: Desarrollo local, levantado por compose.yaml
  - url: https://api.drp.example/api/v1
    description: Entorno ilustrativo (el real se fija al elegir el VPS, al cerrar la Fase 1)
...
security:
  - bearerAuth: []
paths:
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: Listar assets del hogar
      parameters:
        - $ref: "#/components/parameters/LocationIdFilter"
        - $ref: "#/components/parameters/ParentAssetIdFilter"
        - name: ownerId
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/AssetStatus" }
        - name: type
          in: query
          schema: { $ref: "#/components/schemas/AssetType" }
        - name: articleId
          in: query
          schema: { type: string, format: uuid }
        - name: categoryId
          in: query
          schema: { type: string, format: uuid }
        - name: withoutOwner
          in: query
          description: Devuelve los assets huérfanos que dejó una baja de usuario
          schema: { type: boolean, default: false }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: >
            Listado de assets. Excluye los DECOMMISSIONED salvo que se pidan con
            status=DECOMMISSIONED.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssetPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
    post:
      operationId: createAsset
      tags: [assets]
      summary: Dar de alta un asset DURABLE
      description: >
        Solo admite assets DURABLE. La entrada de un CONSUMABLE se hace por
        POST /assets/intake, que resuelve el artículo y suma sobre la existencia
        que ya hubiera en esa ubicación.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssetInput" }
      responses:
        "201":
          description: Asset creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "409":
          description: >
            La ubicación indicada no sirve como contenedor
            (ASSET_LOCATION_CONFLICT): solo un DURABLE puede alojar otros
            assets, porque una estantería contiene cosas y un paquete de harina
            no.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
  /assets/intake:
    post:
      operationId: registerConsumableIntake
      tags: [assets]
      summary: Dar entrada a un consumible
      description: >
        Registra la entrada de un CONSUMABLE. Se indica el artículo (existente
        por articleId, o nuevo mediante el objeto article), la ubicación
        y la cantidad, siempre en la unidad del artículo. Si ya existe una
        existencia de ese artículo en esa ubicación, la operación suma sobre
        ella y responde 200; si no, la crea y responde 201.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssetIntakeInput" }
      responses:
        "200":
          description: >
            Existencia ya presente en esa ubicación: se sumó la cantidad.
            Publica AssetQuantityChanged.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "201":
          description: >
            Primera existencia de ese artículo en esa ubicación. Publica
            AssetCreated, y ArticleCreated si además creó el artículo.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "409":
          description: >
            La ubicación no sirve como contenedor
            (ASSET_LOCATION_CONFLICT), cantidad no positiva
            (INTAKE_QUANTITY_NOT_POSITIVE) o artículo duplicado
            (ARTICLE_DUPLICATE)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
  /assets/{id}:
    parameters:
      - $ref: "#/components/parameters/AssetId"
    get:
      operationId: getAsset
      tags: [assets]
      summary: Detalle de un asset
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "404": { $ref: "#/components/responses/NotFound" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
    patch:
      operationId: updateAsset
      tags: [assets]
      summary: Modificar un asset (ubicación, asset padre o cantidad de un consumible)
      description: >
        La cantidad enviada aquí es absoluta -- sustituye, no suma (para sumar,
        POST /assets/intake). Ajustar cantidad sobre un asset CONSUMABLE publica
        el evento AssetQuantityChanged. El tipo del asset es inmutable; el
        articleId solo admite asignarse a un DURABLE que aún no tenga
        artículo, nunca cambiarse ni retirarse. No acepta estado: la baja se
        hace con DELETE, y LENT/AVAILABLE los gobiernan el préstamo y su
        devolución.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssetPatch" }
      responses:
        "200":
          description: Asset actualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "409":
          description: >
            La ubicación no sirve como contenedor (ASSET_LOCATION_CONFLICT),
            ciclo en la jerarquía (LOCATION_CYCLE), cantidad sobre un asset
            DURABLE (ASSET_QUANTITY_NOT_APPLICABLE), cantidad negativa
            (ASSET_QUANTITY_NEGATIVE) o mover una existencia a una ubicación
            que ya tiene otra viva del mismo artículo
            (EXISTENCE_ALREADY_IN_LOCATION), que se resuelve con
            POST /assets/{id}/merge
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
    delete:
      operationId: decommissionAsset
      tags: [assets]
      summary: Dar de baja un asset
      description: >
        Baja lógica -- pasa el asset a estado DECOMMISSIONED, no borra la fila. Si es una
        existencia con cantidad pendiente, la baja la lleva a 0 y publica
        AssetQuantityChanged antes de AssetDeactivated: lo que quedaba se da
        por perdido. Una existencia dada de baja libera su hueco de artículo y
        ubicación, así que se puede volver a dar entrada allí.
      responses:
        "204":
          description: Asset dado de baja
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: >
            El asset tiene hijos en la jerarquía (ASSET_HAS_CHILDREN) o un
            préstamo ACTIVE (ASSET_HAS_ACTIVE_LOAN)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
  /assets/{id}/merge:
    parameters:
      - $ref: "#/components/parameters/AssetId"
    post:
      operationId: mergeStockItems
      tags: [assets]
      summary: Fusionar dos existencias del mismo artículo
      description: >
        Junta dos existencias CONSUMABLE del mismo artículo creadas por
        separado. El asset de la ruta es el que DESAPARECE: queda a cantidad 0
        y estado DECOMMISSIONED, y su cantidad se suma a la del destino, que conserva su
        ubicación y su propietario. Publica AssetQuantityChanged sobre el
        destino y AssetDeactivated sobre el origen, correlacionados por payload
        (mergedFromAssetId y mergedIntoAssetId).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssetMergeInput" }
      responses:
        "200":
          description: Asset destino con la cantidad acumulada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: >
            Las dos existencias no son del mismo artículo
            (MERGE_ARTICLE_MISMATCH), origen y destino coinciden
            (MERGE_SAME_ASSET), alguna no es CONSUMABLE
            (MERGE_NOT_CONSUMABLE) o alguna está dada de baja
            (MERGE_ASSET_DEACTIVATED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
  /assets/{id}/children:
    parameters:
      - $ref: "#/components/parameters/AssetId"
    get:
      operationId: listAssetChildren
      tags: [assets]
      summary: Hijos directos en la jerarquía de composición
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Listado de assets hijos
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssetPage" }
        "404": { $ref: "#/components/responses/NotFound" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        default: { $ref: "#/components/responses/ServerError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT de usuarios del hogar. Claims -- sub (identityId), memberId
        (pertenencia en el hogar), householdId, role y sid. La distincion entre
        sub y memberId importa: todo lo que el dominio llama "usuario"
        --propietario de un asset, prestador, receptor, createdBy-- apunta a la
        pertenencia, mientras que los refresh tokens cuelgan de la identidad.


        sid identifica cual de las sesiones de esa persona es esta: es el
        identificador del refresh token con el que se emitio. Existe porque
        POST /auth/password revoca las DEMAS sesiones y conserva la que esta en
        uso, y para conservarla hay que saber cual es. La alternativa era pedir
        el refresh token en el cuerpo de esa operacion, que obligaria al cliente
        a mandar una credencial de vida larga donde no hace falta. No es un
        secreto: identifica una fila cuyo valor real esta hasheado.
    loanToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Credencial de un prestador o receptor **externo**, sin cuenta en DRP.
        Es un JWT firmado **sin sub de usuario**, con claims loanId y role
        (LENDER | BORROWER), que viaja como enlace por correo al iniciar el
        prestamo.


        Su alcance son exactamente dos operaciones -- GET /loans/{id} de ESE
        prestamo y POST /loans/{id}/return -- y ninguna otra cosa del hogar.
        Ni el listado, ni otro prestamo, ni los assets.


        Vale a la vez la firma y la fila: su hash se guarda en
        loan_access_tokens, y verificarlo exige firma valida **y** fila viva.
        Las dos capas hacen cosas distintas y por eso estan las dos -- la firma
        lo hace infalsificable y deja rechazar lo caducado sin tocar la base de
        datos; la fila lo hace **revocable**, que un JWT por si solo no puede
...
  parameters:
    AssetId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    LocationIdFilter:
      name: locationId
      in: query
      schema: { type: string, format: uuid }
    ParentAssetIdFilter:
      name: parentAssetId
      in: query
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      description: Pagina solicitada, empezando en 0
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      description: Elementos por pagina
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Unauthorized:
      description: Falta el token, ha caducado o no es válido
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: >
        El cuerpo o los parámetros no cumplen el contrato (VALIDATION_ERROR).
        Es el error de forma; los de negocio van en 409. `details` lleva un
        campo por atributo rechazado.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ServerError:
      description: >
        Error no previsto (INTERNAL_ERROR). El cuerpo nunca lleva detalle
        interno: la traza va al log, no a la respuesta.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      description: >
        Forma única de todo error de la API. `code` es lo que el cliente trata;
        `message` es texto para diagnóstico, no para mostrar tal cual.
      properties:
        code:
          type: string
          example: ASSET_LOCATION_CONFLICT
          description: >
            Enumerado a propósito: el cliente decide qué hacer por el código, y
            un código que no esté aquí es un fallo del contrato, no un caso que
            el frontend deba adivinar.
          enum:
            - ALREADY_MEMBER
            - ARTICLE_DUPLICATE
            - ARTICLE_HAS_EXISTENCES
            - ARTICLE_UNIT_IMMUTABLE
            - ASSET_HAS_ACTIVE_LOAN
            - ASSET_HAS_CHILDREN
            - ASSET_LOCATION_CONFLICT
            - ASSET_QUANTITY_NEGATIVE
            - ASSET_QUANTITY_NOT_APPLICABLE
            - CATEGORY_DUPLICATE
            - CURRENT_PASSWORD_INVALID
            - DOCUMENT_CONTENT_INVALID
            - DOCUMENT_TARGET_INVALID
            - EMAIL_NOT_VERIFIED
            - EXISTENCE_ALREADY_IN_LOCATION
            - FILE_ALREADY_ATTACHED
            - FILE_IN_USE
            - FILE_TOO_LARGE
            - FILE_TYPE_NOT_ALLOWED
            - IDENTITY_ALREADY_MEMBER
            - INTAKE_QUANTITY_NOT_POSITIVE
            - INTERNAL_ERROR
            - INVITATION_ALREADY_PENDING
            - INVITATION_TOKEN_INVALID
            - LOAN_ALREADY_RETURNED
            - LOAN_ASSET_ALREADY_LENT
            - LOAN_ASSET_NOT_DURABLE
            - LOCATION_CYCLE
            - LOCATION_DUPLICATE
            - LOCATION_HAS_ASSETS
            - LOCATION_HAS_CHILDREN
            - MERGE_ARTICLE_MISMATCH
            - MERGE_ASSET_DEACTIVATED
            - MERGE_NOT_CONSUMABLE
            - MERGE_SAME_ASSET
            - MODULE_INACTIVE
            - RATE_LIMITED
            - RESET_TOKEN_INVALID
            - STORAGE_QUOTA_EXCEEDED
            - USER_LAST_ADMIN
            - VALIDATION_ERROR
            - VERIFICATION_TOKEN_INVALID
        message: { type: string }
        details:
          type: object
          description: Contexto adicional del error; su forma depende del código
...
    PageMeta:
      type: object
      required: [page, size, total]
      description: >
        Metadatos de paginación, comunes a todas las colecciones. Son las mismas
        en las doce sin excepción por tamaño esperado: una sola forma que
        aprender en el cliente, y ninguna migración el día que una lista que se
        creía pequeña deje de serlo.
      properties:
        page: { type: integer, minimum: 0, example: 0, description: "Página devuelta, empezando en 0" }
        size: { type: integer, minimum: 1, example: 50, description: Tamaño de página aplicado }
        total: { type: integer, format: int64, example: 137, description: Total de elementos que cumplen el filtro }
...
    AssetStatus:
      type: string
      enum: [AVAILABLE, LENT, DECOMMISSIONED]
    AssetType:
      type: string
      description: >
        Naturaleza del asset. DURABLE tiene identidad propia y se cuenta por
        unidad física; CONSUMABLE se agota o se repone y es una existencia de un
        artículo, con cantidad en la unidad de ese artículo. Se fija en el alta
        y es inmutable después.
      enum: [DURABLE, CONSUMABLE]
...
    AssetInput:
      type: object
      description: >
        Alta de un asset DURABLE. No admite quantity. name y categoryId son
        obligatorios salvo que se informe articleId, en cuyo caso se heredan
        del artículo.
      required: [type]
      properties:
        name: { type: string }
        type: { $ref: "#/components/schemas/AssetType" }
        categoryId: { type: string, format: uuid }
        articleId:
          type: string
          format: uuid
          description: Opcional en un DURABLE -- comparte definición entre unidades idénticas
        ownerId:
          type: string
          format: uuid
          nullable: true
          description: Opcional -- un asset puede quedar sin propietario tras una baja de usuario
        location: { $ref: "#/components/schemas/LocationRef" }
        serialNumber:
          type: string
          description: Solo DURABLE
        acquiredOn:
          type: string
          format: date
          description: Solo DURABLE
        photoUrl:
          type: string
          format: uri
          description: >
            En escritura, enlace externo a la imagen. En lectura, la URL con la
            que se pinta: ese mismo enlace, o la URL FIRMADA de vida corta del
            fichero subido, que caduca en unos 15 minutos y no debe guardarse
            (ver readme.md 5.8.4).
        photoFileId:
          type: string
          format: uuid
          description: Foto subida al servidor. Excluyente con photoUrl
        notes: { type: string }
      # name y categoryId son obligatorios salvo que se herede del articulo.
      anyOf:
        - required: [articleId]
        - required: [name, categoryId]
      # Foto: enlace o fichero, nunca los dos. Los dos vacios SI valen -- no
      # tener foto es lo normal --, asi que esto es "como mucho uno", no oneOf.
      not:
        required: [photoUrl, photoFileId]
    AssetIntakeInput:
      type: object
      description: >
        Entrada de un consumible. Exactamente uno de articleId o article:
        el primero referencia un artículo del catálogo, el segundo lo crea en la
        misma operación. La cantidad va en la unidad del artículo.
      required: [ownerId, quantity]
      properties:
        articleId: { type: string, format: uuid }
        article: { $ref: "#/components/schemas/ArticleInput" }
        ownerId: { type: string, format: uuid }
        location: { $ref: "#/components/schemas/LocationRef" }
        quantity:
          type: number
          exclusiveMinimum: true
          minimum: 0
          description: Cantidad que entra, se suma a la existencia si ya la hay
      oneOf:
        - required: [articleId]
        - required: [article]
    AssetMergeInput:
      type: object
      description: >
        Destino de la fusión. El origen es el asset de la ruta, que es el que
        desaparece.
      required: [targetAssetId]
      properties:
        targetAssetId:
          type: string
          format: uuid
          description: Existencia que absorbe la cantidad y sobrevive
    AssetPatch:
      type: object
      description: >
        Modificación parcial. El tipo es inmutable y no se acepta aquí.
        cantidad solo es válida sobre un asset CONSUMABLE, y es absoluta.
        nombre y categoryId solo son válidos en un asset sin artículo -- si lo
        tiene, se cambian en el artículo. estado no se acepta: lo gobiernan
        DELETE /assets/{id}, POST /loans y POST /loans/{id}/return. ownerId es
        la vía para reasignar un asset huérfano. serialNumber y acquiredOn son
        al revés que cantidad: solo valen sobre un DURABLE.
      properties:
        name: { type: string }
        categoryId: { type: string, format: uuid }
        articleId:
          type: string
          format: uuid
          description: Solo para asignar artículo a un DURABLE que aún no lo tiene
        ownerId: { type: string, format: uuid }
        location: { $ref: "#/components/schemas/LocationRef" }
        quantity: { type: number, minimum: 0 }
        serialNumber:
          type: string
          nullable: true
          description: >
            Solo sobre un DURABLE. Se corrige después del alta porque casi nunca
            se tiene delante al dar de alta: la etiqueta está detrás del aparato.
        acquiredOn:
          type: string
          format: date
          nullable: true
          description: Solo sobre un DURABLE. Es procedencia, no valor.
        photoUrl: { type: string, format: uri, nullable: true }
        photoFileId:
          type: string
          format: uuid
          nullable: true
          description: >
            Excluyente con photoUrl. Ponerlo a null desadjunta la foto y marca
            el fichero para su retirada: la cuota se libera en el acto y los
            bytes los desenlaza el proceso diario.
        notes: { type: string }
    Asset:
      allOf:
        - type: object
          properties:
            id: { type: string, format: uuid }
            name:
              type: string
              description: Propio del asset, o resuelto desde su artículo cuando lo tiene.
            type: { $ref: "#/components/schemas/AssetType" }
            categoryId:
              type: string
              format: uuid
              description: Propia del asset, o resuelta desde su artículo cuando lo tiene.
            category:
              type: string
              description: Nombre de la categoría, resuelto para lectura.
            articleId:
              type: string
              format: uuid
              nullable: true
              description: Siempre presente en CONSUMABLE; opcional en DURABLE.
            ownerId:
              type: string
              format: uuid
              nullable: true
              description: Nulo si su propietario causó baja y aún no se ha reasignado.
            serialNumber: { type: string, nullable: true }
            acquiredOn: { type: string, format: date, nullable: true }
            photoUrl:
              type: string
              format: uri
              nullable: true
              description: >
                Enlace externo, o URL FIRMADA de vida corta cuando la foto vive
                aqui. Caduca con el access token y no debe guardarse.
            photoThumbnailUrl:
              type: string
              format: uri
              nullable: true
              description: >
                Miniatura firmada, para listados. Nula si la foto es un enlace
                externo, porque de esa no se genera nada.
            photoFileId: { type: string, format: uuid, nullable: true }
            notes: { type: string, nullable: true }
            location: { $ref: "#/components/schemas/LocationRef" }
            status: { $ref: "#/components/schemas/AssetStatus" }
            quantity:
              type: number
              nullable: true
              description: Solo presente en assets CONSUMABLE.
            unit:
              allOf:
                - $ref: "#/components/schemas/MeasurementUnit"
              nullable: true
              description: Resuelta desde el artículo; solo presente si el asset lo tiene.
            warnings:
              type: array
              description: >
                Advertencias de la operación que acaba de tener éxito. Vacío en
                el caso normal. NO son errores: la operación se hizo, la
                respuesta es 200 o 201, y el asset devuelto es el estado real.
                Existe porque hay una regla que advierte sin bloquear -- superar
                la capacidad declarada de una ubicación (ver readme.md 4.1.2) --
                y sin este campo no habría por dónde contarlo: el sistema no
                sabe cuánto ocupa cada cosa, así que bloquear con datos
                incompletos impediría guardar algo que sí cabe.
              items: { $ref: "#/components/schemas/Warning" }
            createdAt: { type: string, format: date-time }
        - $ref: "#/components/schemas/Authorship"
...
    AssetPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/Asset" }
...